requirements.txt
# Portugal Active — Master Data import scripts
# Python 3.11+ recommended (works on 3.9+)

psycopg[binary]>=3.1,<4
openpyxl>=3.1,<4
python-dotenv>=1.0,<2
unidecode>=1.3,<2
